Add stats

+++ GGJam 2021/GameObject/Portrait.cs
@@ -12,7 +12,7 @@ namespace GGJam_2021 {
 
         private int paranoia;
 
-        public Portrait(string textureName, LayerMask layerMask, Scene scene, Scene nextScene, ColliderType colliderType, int fotogrammi = 0, int fps = 0, int w = 0, int paranoia = 0) : base(textureName, layerMask, scene, nextScene, colliderType, w, 0) {
+        public Portrait(string textureName, LayerMask layerMask, Scene scene, Scene nextScene, ColliderType colliderType, int fotogrammi = 0, int fps = 0, int w = 0, float paranoia = 0) : base(textureName, layerMask, scene, nextScene, colliderType, w, 0) {
             textureOffset = Vector2.Zero;
             isOpened = false;
             if (fotogrammi == 0) {
